# generated/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "IApiValidationError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["field", "message"],
        "type": "object",
        "additionalProperties": false
      },
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
      "TErrorCause": {
        "$ref": "#/components/schemas/Record_string.unknown_"
      },
      "IHttpErrorDto": {
        "properties": {
          "name": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          },
          "validationErrors": {
            "items": {
              "$ref": "#/components/schemas/IApiValidationError"
            },
            "type": "array"
          },
          "cause": {
            "$ref": "#/components/schemas/TErrorCause"
          }
        },
        "required": ["name", "errorKey"],
        "type": "object",
        "additionalProperties": false
      },
...
      "IUserSignupReq": {
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["firstName", "lastName", "email", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "IAccessToken": {
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": ["accessToken"],
        "type": "object",
        "additionalProperties": false
      },
      "IEmailLoginReq": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "IResetPasswordReq": {
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          }
        },
        "required": ["token", "password", "confirmPassword"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "purple-ledger-core",
    "version": "1.0.0",
    "description": "This repo contains the server-side source code for Purple Ledger's bookkeeping and tax app.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "contact": {
      "name": "Osahon Oboite"
    }
  },
  "paths": {
...
    "/auth/signup-with-email": {
      "post": {
        "operationId": "signupWithEmail",
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "User signup with email and password",
        "tags": ["Auth"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IUserSignupReq"
              }
            }
          }
        }
      }
    },
    "/auth/signup/complete": {
      "post": {
        "operationId": "verifyEmail",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IAccessToken"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "\nVerify user email",
        "tags": ["Auth"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/login-with-email": {
      "post": {
        "operationId": "loginWithEmail",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IAccessToken"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "User login with email and password",
        "tags": ["Auth"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IEmailLoginReq"
              }
            }
          }
        }
      }
    },
    "/auth/get-password-reset-link": {
      "post": {
        "operationId": "getPasswordResetLink",
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "\nGet password reset link",
        "tags": ["Auth"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["email"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IAccessToken"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "\nReset password",
        "tags": ["Auth"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IResetPasswordReq"
              }
            }
          }
        }
      }
    },
...
    "/auth/refresh-access-token": {
      "post": {
        "operationId": "refreshAccessToken",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IAccessToken"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Refresh access token",
        "tags": ["Auth"],
        "security": [],
        "parameters": []
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}